Refactor map components to be reused
